pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt2-ivr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "pandas~=2.2.0",
    "PyYAML~=6.0",
    "protobuf>=6.33.5",
    "pyarrow>=23.0.0",
    "pyfiglet>=1.0.4",
    "pyright>=1.1.408",
    "tokenizers>=0.22.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "datasets>=4.5.0",
    "rich>=14.3.2",
]

[project.scripts]
ivr = "gpt2_ivr.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pyright]
typeCheckingMode = "basic"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[dependency-groups]
dev = [
    "black>=26.1.0",
]
